shop_user_service/requirements.txt
asgiref
Django
djangorestframework
Pillow
pytz
sqlparse
typing_extensions
pytest
pytest-django
requests
django-cors-headers
coreapi
coreschema
dj-rest-auth
django-allauth
djangorestframework-simplejwt
django-filter
httpie
mysqlclient